refactor(drizzle-utils): split snapshotSchema by dialect
The 1003-line snapshotSchema.ts is now decomposed into a snapshot-schema/
folder organised by dialect:

- types.ts — public snapshot types
- shared.ts — cross-dialect helpers (quoting, coercion, etc.)
- postgres.ts — PostgreSQL + CockroachDB loaders (snapshotPostgres)
- mysql.ts — MySQL loaders (snapshotMysql)
- snapshotSchema.ts — dispatcher
- index.ts — barrel exposing snapshotSchema + types

No behavior changes; tsc and the unit-level diff tests still pass.
Internal importers updated to the new path.
