+++ packages/app/drizzle-utils/src/snapshot-schema/types.ts
@@ -1,0 +1,112 @@
+import type { SqlExecutor } from '../markMigrationsApplied.ts'
+
+export type SnapshotDialect = 'postgresql' | 'mysql' | 'cockroachdb'
+
+export interface ColumnSnapshot {
+  /** Canonical SQL type as reported by the database catalog (e.g. "integer", "character varying(255)"). */
+  type: string
+  nullable: boolean
+  /** Raw default expression as stored in the catalog, or null if there is no default. */
+  default: string | null
+  /**
+   * Identity column kind (PostgreSQL/CockroachDB only):
+   * - `'always'` — `GENERATED ALWAYS AS IDENTITY`
+   * - `'by-default'` — `GENERATED BY DEFAULT AS IDENTITY`
+   * - `null` — not an identity column (may still be a `SERIAL` — that surfaces as a sequence-based default).
+   */
+  identity: 'always' | 'by-default' | null
+  /** MySQL "extra" attributes such as "auto_increment", "on update CURRENT_TIMESTAMP". Empty string when absent. */
+  extra?: string
+  /** Generated/computed column expression, if any. */
+  generated: string | null
+}
+
+export interface PrimaryKeySnapshot {
+  name: string
+  columns: string[]
+}
+
+export interface UniqueConstraintSnapshot {
+  columns: string[]
+  /** For PostgreSQL, whether the constraint is DEFERRABLE. */
+  deferrable?: boolean
+}
+
+export type ForeignKeyAction = 'NO ACTION' | 'RESTRICT' | 'CASCADE' | 'SET NULL' | 'SET DEFAULT'
+
+export interface ForeignKeySnapshot {
+  columns: string[]
+  referencedSchema: string
+  referencedTable: string
+  referencedColumns: string[]
+  onUpdate: ForeignKeyAction
+  onDelete: ForeignKeyAction
+}
+
+export interface CheckConstraintSnapshot {
+  /** Normalized check expression as reported by the catalog. */
+  expression: string
+}
+
+export interface IndexSnapshot {
+  columns: string[]
+  unique: boolean
+  /**
+   * Index method (PostgreSQL): "btree", "gin", "hash", etc.
+   * Always "BTREE" / "HASH" / etc. on MySQL.
+   */
+  method: string
+  /** Partial-index predicate (PostgreSQL only). */
+  predicate?: string | null
+}
+
+export interface TableSnapshot {
+  columns: Record<string, ColumnSnapshot>
+  primaryKey: PrimaryKeySnapshot | null
+  uniqueConstraints: Record<string, UniqueConstraintSnapshot>
+  foreignKeys: Record<string, ForeignKeySnapshot>
+  checkConstraints: Record<string, CheckConstraintSnapshot>
+  indexes: Record<string, IndexSnapshot>
+}
+
+export interface EnumSnapshot {
+  values: string[]
+}
+
+export interface SequenceSnapshot {
+  dataType: string
+  start: string
+  increment: string
+  minValue: string
+  maxValue: string
+  cycle: boolean
+}
+
+export interface SchemaContents {
+  tables: Record<string, TableSnapshot>
+  enums: Record<string, EnumSnapshot>
+  sequences: Record<string, SequenceSnapshot>
+}
+
+export interface SchemaSnapshot {
+  dialect: SnapshotDialect
+  schemas: Record<string, SchemaContents>
+}
+
+export interface SnapshotSchemaOptions {
+  executor: SqlExecutor
+  dialect: SnapshotDialect
+  /**
+   * Schemas to capture.
+   * - PostgreSQL / CockroachDB: defaults to `['public']`.
+   * - MySQL: defaults to the current database (queried via `SELECT DATABASE()`).
+   */
+  schemas?: string[]
+  /**
+   * Tables to exclude from the snapshot.
+   * Matches by unqualified table name. Defaults to common migration-tracking tables
+   * (`__drizzle_migrations`, `_prisma_migrations`) so the source and target snapshots
+   * align after a Prisma → Drizzle baseline.
+   */
+  excludeTables?: string[]
+}
