--- packages/app/drizzle-utils/src/snapshotSchema.ts
@@ -1,1003 +1,0 @@
-import type { SqlExecutor } from './markMigrationsApplied.ts'
-
-export type SnapshotDialect = 'postgresql' | 'mysql' | 'cockroachdb'
-
-export interface ColumnSnapshot {
-  /** Canonical SQL type as reported by the database catalog (e.g. "integer", "character varying(255)"). */
-  type: string
-  nullable: boolean
-  /** Raw default expression as stored in the catalog, or null if there is no default. */
-  default: string | null
-  /**
-   * Identity column kind (PostgreSQL/CockroachDB only):
-   * - `'always'` — `GENERATED ALWAYS AS IDENTITY`
-   * - `'by-default'` — `GENERATED BY DEFAULT AS IDENTITY`
-   * - `null` — not an identity column (may still be a `SERIAL` — that surfaces as a sequence-based default).
-   */
-  identity: 'always' | 'by-default' | null
-  /** MySQL "extra" attributes such as "auto_increment", "on update CURRENT_TIMESTAMP". Empty string when absent. */
-  extra?: string
-  /** Generated/computed column expression, if any. */
-  generated: string | null
-}
-
-export interface PrimaryKeySnapshot {
-  name: string
-  columns: string[]
-}
-
-export interface UniqueConstraintSnapshot {
-  columns: string[]
-  /** For PostgreSQL, whether the constraint is DEFERRABLE. */
-  deferrable?: boolean
-}
-
-export type ForeignKeyAction = 'NO ACTION' | 'RESTRICT' | 'CASCADE' | 'SET NULL' | 'SET DEFAULT'
-
-export interface ForeignKeySnapshot {
-  columns: string[]
-  referencedSchema: string
-  referencedTable: string
-  referencedColumns: string[]
-  onUpdate: ForeignKeyAction
-  onDelete: ForeignKeyAction
-}
-
-export interface CheckConstraintSnapshot {
-  /** Normalized check expression as reported by the catalog. */
-  expression: string
-}
-
-export interface IndexSnapshot {
-  columns: string[]
-  unique: boolean
-  /**
-   * Index method (PostgreSQL): "btree", "gin", "hash", etc.
-   * Always "BTREE" / "HASH" / etc. on MySQL.
-   */
-  method: string
-  /** Partial-index predicate (PostgreSQL only). */
-  predicate?: string | null
-}
-
-export interface TableSnapshot {
-  columns: Record<string, ColumnSnapshot>
-  primaryKey: PrimaryKeySnapshot | null
-  uniqueConstraints: Record<string, UniqueConstraintSnapshot>
-  foreignKeys: Record<string, ForeignKeySnapshot>
-  checkConstraints: Record<string, CheckConstraintSnapshot>
-  indexes: Record<string, IndexSnapshot>
-}
-
-export interface EnumSnapshot {
-  values: string[]
-}
-
-export interface SequenceSnapshot {
-  dataType: string
-  start: string
-  increment: string
-  minValue: string
-  maxValue: string
-  cycle: boolean
-}
-
-export interface SchemaContents {
-  tables: Record<string, TableSnapshot>
-  enums: Record<string, EnumSnapshot>
-  sequences: Record<string, SequenceSnapshot>
-}
-
-export interface SchemaSnapshot {
-  dialect: SnapshotDialect
-  schemas: Record<string, SchemaContents>
-}
-
-export interface SnapshotSchemaOptions {
-  executor: SqlExecutor
-  dialect: SnapshotDialect
-  /**
-   * Schemas to capture.
-   * - PostgreSQL / CockroachDB: defaults to `['public']`.
-   * - MySQL: defaults to the current database (queried via `SELECT DATABASE()`).
-   */
-  schemas?: string[]
-  /**
-   * Tables to exclude from the snapshot.
-   * Matches by unqualified table name. Defaults to common migration-tracking tables
-   * (`__drizzle_migrations`, `_prisma_migrations`) so the source and target snapshots
-   * align after a Prisma → Drizzle baseline.
-   */
-  excludeTables?: string[]
-}
-
-const DEFAULT_EXCLUDE_TABLES = ['__drizzle_migrations', '_prisma_migrations']
-
-function isPgLike(dialect: SnapshotDialect): boolean {
-  return dialect === 'postgresql' || dialect === 'cockroachdb'
-}
-
-function emptySchemaContents(): SchemaContents {
-  return { tables: {}, enums: {}, sequences: {} }
-}
-
-function emptyTable(): TableSnapshot {
-  return {
-    columns: {},
-    primaryKey: null,
-    uniqueConstraints: {},
-    foreignKeys: {},
-    checkConstraints: {},
-    indexes: {},
-  }
-}
-
-function quotePgLiteralList(values: string[]): string {
-  if (values.length === 0) return "''"
-  return values.map((v) => `'${v.replace(/'/g, "''")}'`).join(', ')
-}
-
-function quoteMysqlLiteralList(values: string[]): string {
-  if (values.length === 0) return "''"
-  return values.map((v) => `'${v.replace(/'/g, "''").replace(/\\/g, '\\\\')}'`).join(', ')
-}
-
-function asString(value: unknown): string {
-  if (value === null || value === undefined) return ''
-  return String(value)
-}
-
-function asNullableString(value: unknown): string | null {
-  if (value === null || value === undefined) return null
-  return String(value)
-}
-
-function asBool(value: unknown): boolean {
-  if (typeof value === 'boolean') return value
-  if (value === 1 || value === '1') return true
-  if (value === 0 || value === '0') return false
-  if (value === 'YES' || value === 'yes' || value === 't' || value === 'true') return true
-  return false
-}
-
-function ensureTable(contents: SchemaContents, tableName: string): TableSnapshot {
-  let table = contents.tables[tableName]
-  if (!table) {
-    table = emptyTable()
-    contents.tables[tableName] = table
-  }
-  return table
-}
-
-function normalizeAction(action: string | null | undefined): ForeignKeyAction {
-  if (!action) return 'NO ACTION'
-  const upper = action.toUpperCase()
-  switch (upper) {
-    case 'NO ACTION':
-    case 'RESTRICT':
-    case 'CASCADE':
-    case 'SET NULL':
-    case 'SET DEFAULT':
-      return upper
-    default:
-      return 'NO ACTION'
-  }
-}
-
-// ──────────────────── PostgreSQL ────────────────────
-
-async function snapshotPostgres(
-  executor: SqlExecutor,
-  schemas: string[],
-  excludeTables: Set<string>,
-  dialect: SnapshotDialect,
-): Promise<SchemaSnapshot> {
-  const snapshot: SchemaSnapshot = { dialect, schemas: {} }
-  for (const schema of schemas) {
-    snapshot.schemas[schema] = emptySchemaContents()
-  }
-
-  const schemaList = quotePgLiteralList(schemas)
-  const excludeList = excludeTables.size > 0 ? quotePgLiteralList([...excludeTables]) : "''"
-
-  await loadPgTables(executor, snapshot, schemaList, excludeList)
-  await loadPgColumns(executor, snapshot, schemaList, excludeList)
-  // attMap is shared between constraint and index loading — fetch once.
-  const attMap = await loadPgAttMap(executor, schemaList, excludeList)
-  await loadPgConstraints(executor, snapshot, schemaList, excludeList, attMap)
-  await loadPgIndexes(executor, snapshot, schemaList, excludeList, attMap)
-  await loadPgEnums(executor, snapshot, schemaList)
-  await loadPgSequences(executor, snapshot, schemaList)
-
-  return snapshot
-}
-
-async function loadPgTables(
-  executor: SqlExecutor,
-  snapshot: SchemaSnapshot,
-  schemaList: string,
-  excludeList: string,
-): Promise<void> {
-  const tables = await executor.all(
-    `SELECT table_schema, table_name
-     FROM information_schema.tables
-     WHERE table_schema IN (${schemaList})
-       AND table_type = 'BASE TABLE'
-       AND table_name NOT IN (${excludeList})
-     ORDER BY table_schema, table_name`,
-  )
-  for (const row of tables) {
-    const schema = asString(row.table_schema)
-    const name = asString(row.table_name)
-    const contents = snapshot.schemas[schema]
-    if (!contents) continue
-    ensureTable(contents, name)
-  }
-}
-
-async function loadPgColumns(
-  executor: SqlExecutor,
-  snapshot: SchemaSnapshot,
-  schemaList: string,
-  excludeList: string,
-): Promise<void> {
-  const columns = await executor.all(
-    `SELECT
-        c.table_schema,
-        c.table_name,
-        c.column_name,
-        c.data_type,
-        c.udt_schema,
-        c.udt_name,
-        c.character_maximum_length,
-        c.numeric_precision,
-        c.numeric_scale,
-        c.datetime_precision,
-        c.is_nullable,
-        c.column_default,
-        c.is_identity,
-        c.identity_generation,
-        c.is_generated,
-        c.generation_expression
-      FROM information_schema.columns c
-      WHERE c.table_schema IN (${schemaList})
-        AND c.table_name NOT IN (${excludeList})
-      ORDER BY c.table_schema, c.table_name, c.ordinal_position`,
-  )
-
-  for (const row of columns) {
-    const schema = asString(row.table_schema)
-    const tableName = asString(row.table_name)
-    const contents = snapshot.schemas[schema]
-    if (!contents) continue
-    const table = contents.tables[tableName]
-    if (!table) continue
-    table.columns[asString(row.column_name)] = buildPgColumn(row)
-  }
-}
-
-function buildPgColumn(row: Record<string, unknown>): ColumnSnapshot {
-  return {
-    type: buildPgType(row),
-    nullable: asString(row.is_nullable) === 'YES',
-    default: asNullableString(row.column_default),
-    identity: parsePgIdentity(row),
-    generated:
-      asString(row.is_generated) === 'ALWAYS' ? asNullableString(row.generation_expression) : null,
-  }
-}
-
-function parsePgIdentity(row: Record<string, unknown>): 'always' | 'by-default' | null {
-  if (asString(row.is_identity) !== 'YES') return null
-  return asString(row.identity_generation).toUpperCase() === 'ALWAYS' ? 'always' : 'by-default'
-}
-
-function buildPgType(row: Record<string, unknown>): string {
-  const dataType = asString(row.data_type)
-  return (
-    buildPgUserDefinedType(dataType, row) ??
-    buildPgCharType(dataType, row.character_maximum_length) ??
-    buildPgNumericType(dataType, row.numeric_precision, row.numeric_scale) ??
-    buildPgDatetimeType(dataType, row.datetime_precision) ??
-    dataType
-  )
-}
-
-function buildPgUserDefinedType(dataType: string, row: Record<string, unknown>): string | null {
-  if (dataType !== 'USER-DEFINED' && dataType !== 'ARRAY') return null
-  const udt = asString(row.udt_name)
-  if (!udt) return null
-  // Qualify user-defined types with their schema so two enums of the same name
-  // in different schemas don't snapshot as identical. Built-ins (pg_catalog) stay unqualified
-  // to keep names like `integer` / `_int4` readable.
-  const udtSchema = asString(row.udt_schema)
-  return udtSchema && udtSchema !== 'pg_catalog' ? `${udtSchema}.${udt}` : udt
-}
-
-function buildPgCharType(dataType: string, charLen: unknown): string | null {
-  if (charLen == null) return null
-  if (dataType === 'character varying') return `character varying(${charLen})`
-  if (dataType === 'character') return `character(${charLen})`
-  if (dataType === 'bit') return `bit(${charLen})`
-  if (dataType === 'bit varying') return `bit varying(${charLen})`
-  return null
-}
-
-function buildPgNumericType(
-  dataType: string,
-  numPrecision: unknown,
-  numScale: unknown,
-): string | null {
-  if (dataType !== 'numeric' || numPrecision == null) return null
-  if (numScale != null) return `numeric(${numPrecision},${numScale})`
-  return `numeric(${numPrecision})`
-}
-
-function buildPgDatetimeType(dataType: string, datetimePrecision: unknown): string | null {
-  if (datetimePrecision == null) return null
-  // time/timestamp/interval may carry a fractional-seconds precision. information_schema
-  // returns data_type as e.g. "timestamp without time zone" without baking precision into it,
-  // so inject it after the base word: "timestamp(3) without time zone".
-  // Order matters: `timestamp` must come before `time` so the longer prefix wins.
-  const match = /^(timestamp|time|interval)(.*)$/.exec(dataType)
-  if (!match) return null
-  return `${match[1]}(${datetimePrecision})${match[2]}`
-}
-
-async function loadPgAttMap(
-  executor: SqlExecutor,
-  schemaList: string,
-  excludeList: string,
-): Promise<Map<string, Map<number, string>>> {
-  const attRows = await executor.all(
-    `SELECT n.nspname AS schema_name, cl.relname AS table_name, a.attnum, a.attname
-      FROM pg_attribute a
-      JOIN pg_class cl ON cl.oid = a.attrelid
-      JOIN pg_namespace n ON n.oid = cl.relnamespace
-      WHERE n.nspname IN (${schemaList})
-        AND cl.relname NOT IN (${excludeList})
-        AND a.attnum > 0
-        AND NOT a.attisdropped`,
-  )
-
-  const attMap = new Map<string, Map<number, string>>()
-  for (const row of attRows) {
-    const key = `${asString(row.schema_name)}.${asString(row.table_name)}`
-    let inner = attMap.get(key)
-    if (!inner) {
-      inner = new Map<number, string>()
-      attMap.set(key, inner)
-    }
-    inner.set(Number(row.attnum), asString(row.attname))
-  }
-  return attMap
-}
-
-function resolvePgColumns(
-  attMap: Map<string, Map<number, string>>,
-  schemaName: string,
-  tableName: string,
-  attnums: unknown,
-): string[] {
-  const inner = attMap.get(`${schemaName}.${tableName}`)
-  if (!inner) return []
-  const arr = parsePgIntArray(attnums)
-  return arr.map((n) => inner.get(n) ?? `attnum:${n}`)
-}
-
-async function loadPgConstraints(
-  executor: SqlExecutor,
-  snapshot: SchemaSnapshot,
-  schemaList: string,
-  excludeList: string,
-  attMap: Map<string, Map<number, string>>,
-): Promise<void> {
-  // pg_constraint gives us the canonical view across types.
-  // contype: 'p' primary, 'u' unique, 'f' foreign, 'c' check.
-  const constraintRows = await executor.all(
-    `SELECT
-        n.nspname AS schema_name,
-        cl.relname AS table_name,
-        c.conname AS constraint_name,
-        c.contype::text AS contype,
-        c.condeferrable AS deferrable,
-        c.confupdtype::text AS confupdtype,
-        c.confdeltype::text AS confdeltype,
-        pg_get_constraintdef(c.oid) AS definition,
-        c.conkey AS conkey,
-        c.confkey AS confkey,
-        rn.nspname AS ref_schema,
-        rcl.relname AS ref_table
-      FROM pg_constraint c
-      JOIN pg_namespace n ON n.oid = c.connamespace
-      JOIN pg_class cl ON cl.oid = c.conrelid
-      LEFT JOIN pg_class rcl ON rcl.oid = c.confrelid
-      LEFT JOIN pg_namespace rn ON rn.oid = rcl.relnamespace
-      WHERE n.nspname IN (${schemaList})
-        AND cl.relname NOT IN (${excludeList})
-        AND c.contype IN ('p', 'u', 'f', 'c')
-      ORDER BY n.nspname, cl.relname, c.conname`,
-  )
-
-  for (const row of constraintRows) {
-    const schemaName = asString(row.schema_name)
-    const tableName = asString(row.table_name)
-    const contents = snapshot.schemas[schemaName]
-    if (!contents) continue
-    const table = contents.tables[tableName]
-    if (!table) continue
-    applyPgConstraintRow(table, row, schemaName, tableName, attMap)
-  }
-}
-
-function applyPgConstraintRow(
-  table: TableSnapshot,
-  row: Record<string, unknown>,
-  schemaName: string,
-  tableName: string,
-  attMap: Map<string, Map<number, string>>,
-): void {
-  const constraintName = asString(row.constraint_name)
-  const contype = asString(row.contype)
-  const cols = (a: unknown) => resolvePgColumns(attMap, schemaName, tableName, a)
-
-  if (contype === 'p') {
-    table.primaryKey = { name: constraintName, columns: cols(row.conkey) }
-    return
-  }
-  if (contype === 'u') {
-    table.uniqueConstraints[constraintName] = {
-      columns: cols(row.conkey),
-      deferrable: asBool(row.deferrable),
-    }
-    return
-  }
-  if (contype === 'f') {
-    const refSchema = asString(row.ref_schema)
-    const refTable = asString(row.ref_table)
-    table.foreignKeys[constraintName] = {
-      columns: cols(row.conkey),
-      referencedSchema: refSchema,
-      referencedTable: refTable,
-      referencedColumns: resolvePgColumns(attMap, refSchema, refTable, row.confkey),
-      onUpdate: pgFkActionFromCode(asString(row.confupdtype)),
-      onDelete: pgFkActionFromCode(asString(row.confdeltype)),
-    }
-    return
-  }
-  if (contype === 'c') {
-    table.checkConstraints[constraintName] = {
-      expression: normalizeCheckDefinition(asString(row.definition)),
-    }
-  }
-}
-
-function pgFkActionFromCode(code: string): ForeignKeyAction {
-  switch (code) {
-    case 'a':
-      return 'NO ACTION'
-    case 'r':
-      return 'RESTRICT'
-    case 'c':
-      return 'CASCADE'
-    case 'n':
-      return 'SET NULL'
-    case 'd':
-      return 'SET DEFAULT'
-    default:
-      return 'NO ACTION'
-  }
-}
-
-function parsePgIntArray(value: unknown): number[] {
-  if (Array.isArray(value)) {
-    return value.map((v) => Number(v)).filter((n) => Number.isFinite(n))
-  }
-  if (typeof value === 'string') {
-    const cleaned = value.replace(/^\{|\}$/g, '').trim()
-    if (!cleaned) return []
-    return cleaned
-      .split(',')
-      .map((s) => Number(s.trim()))
-      .filter((n) => Number.isFinite(n))
-  }
-  return []
-}
-
-function normalizeCheckDefinition(raw: string): string {
-  // pg_get_constraintdef returns 'CHECK ((expr))' — strip leading 'CHECK ' and surrounding parens
-  // so two databases storing the same check don't diff because of redundant outer parens.
-  const trimmed = raw.replace(/^CHECK\s*/i, '').trim()
-  return trimmed
-}
-
-async function loadPgIndexes(
-  executor: SqlExecutor,
-  snapshot: SchemaSnapshot,
-  schemaList: string,
-  excludeList: string,
-  attMap: Map<string, Map<number, string>>,
-): Promise<void> {
-  // pg_get_indexdef(index, column, pretty) returns the per-column expression text. We aggregate
-  // them with a unit-separator delimiter (CHR(31)) so we can split on something that can't appear
-  // in identifiers or expression syntax, then use the n=0 (expression column) entries by position.
-  const rows = await executor.all(
-    `SELECT
-        n.nspname AS schema_name,
-        cl.relname AS table_name,
-        ic.relname AS index_name,
-        am.amname AS method,
-        ix.indisunique AS is_unique,
-        ix.indkey::text AS indkey,
-        pg_get_expr(ix.indpred, ix.indrelid) AS predicate,
-        COALESCE(
-          (SELECT string_agg(pg_get_indexdef(ix.indexrelid, k::int, false), CHR(31) ORDER BY k)
-           FROM generate_subscripts(ix.indkey, 1) AS k),
-          ''
-        ) AS index_exprs
-      FROM pg_index ix
-      JOIN pg_class ic ON ic.oid = ix.indexrelid
-      JOIN pg_class cl ON cl.oid = ix.indrelid
-      JOIN pg_namespace n ON n.oid = cl.relnamespace
-      JOIN pg_am am ON am.oid = ic.relam
-      WHERE n.nspname IN (${schemaList})
-        AND cl.relname NOT IN (${excludeList})
-      ORDER BY n.nspname, cl.relname, ic.relname`,
-  )
-
-  for (const row of rows) {
-    const schemaName = asString(row.schema_name)
-    const tableName = asString(row.table_name)
-    const contents = snapshot.schemas[schemaName]
-    if (!contents) continue
-    const table = contents.tables[tableName]
-    if (!table) continue
-
-    const indexName = asString(row.index_name)
-    const inner = attMap.get(`${schemaName}.${tableName}`)
-    const indkey = parsePgIntArray(row.indkey)
-    const exprText = asString(row.index_exprs)
-    const exprs = exprText === '' ? [] : exprText.split('\x1f')
-    const columns = indkey.map((n, idx) => {
-      if (n === 0) {
-        // Expression index column: use the per-position pg_get_indexdef text so that
-        // lower(email) and upper(email) snapshot distinctly.
-        return exprs[idx] ?? '<expression>'
-      }
-      return inner?.get(n) ?? `attnum:${n}`
-    })
-
-    table.indexes[indexName] = {
-      columns,
-      unique: asBool(row.is_unique),
-      method: asString(row.method),
-      predicate: asNullableString(row.predicate),
-    }
-  }
-}
-
-async function loadPgEnums(
-  executor: SqlExecutor,
-  snapshot: SchemaSnapshot,
-  schemaList: string,
-): Promise<void> {
-  const rows = await executor.all(
-    `SELECT n.nspname AS schema_name, t.typname AS enum_name, e.enumlabel AS value, e.enumsortorder
-      FROM pg_type t
-      JOIN pg_enum e ON e.enumtypid = t.oid
-      JOIN pg_namespace n ON n.oid = t.typnamespace
-      WHERE n.nspname IN (${schemaList})
-      ORDER BY n.nspname, t.typname, e.enumsortorder`,
-  )
-  for (const row of rows) {
-    const schemaName = asString(row.schema_name)
-    const enumName = asString(row.enum_name)
-    const contents = snapshot.schemas[schemaName]
-    if (!contents) continue
-    let entry = contents.enums[enumName]
-    if (!entry) {
-      entry = { values: [] }
-      contents.enums[enumName] = entry
-    }
-    entry.values.push(asString(row.value))
-  }
-}
-
-async function loadPgSequences(
-  executor: SqlExecutor,
-  snapshot: SchemaSnapshot,
-  schemaList: string,
-): Promise<void> {
-  const rows = await executor.all(
-    `SELECT
-        sequence_schema,
-        sequence_name,
-        data_type,
-        start_value,
-        increment,
-        minimum_value,
-        maximum_value,
-        cycle_option
-      FROM information_schema.sequences
-      WHERE sequence_schema IN (${schemaList})
-      ORDER BY sequence_schema, sequence_name`,
-  )
-  for (const row of rows) {
-    const schemaName = asString(row.sequence_schema)
-    const name = asString(row.sequence_name)
-    const contents = snapshot.schemas[schemaName]
-    if (!contents) continue
-    contents.sequences[name] = {
-      dataType: asString(row.data_type),
-      start: asString(row.start_value),
-      increment: asString(row.increment),
-      minValue: asString(row.minimum_value),
-      maxValue: asString(row.maximum_value),
-      cycle: asString(row.cycle_option).toUpperCase() === 'YES',
-    }
-  }
-}
-
-// ──────────────────── MySQL ────────────────────
-
-async function snapshotMysql(
-  executor: SqlExecutor,
-  schemas: string[] | undefined,
-  excludeTables: Set<string>,
-): Promise<SchemaSnapshot> {
-  const schemaList = await resolveMysqlSchemaList(executor, schemas)
-
-  const snapshot: SchemaSnapshot = { dialect: 'mysql', schemas: {} }
-  for (const schema of schemaList) {
-    snapshot.schemas[schema] = emptySchemaContents()
-  }
-
-  const schemasIn = quoteMysqlLiteralList(schemaList)
-  const excludeIn = excludeTables.size > 0 ? quoteMysqlLiteralList([...excludeTables]) : "''"
-
-  await loadMysqlTables(executor, snapshot, schemasIn, excludeIn)
-  await loadMysqlColumns(executor, snapshot, schemasIn, excludeIn)
-  await loadMysqlIndexesAndKeys(executor, snapshot, schemasIn, excludeIn)
-  await loadMysqlForeignKeys(executor, snapshot, schemasIn, excludeIn)
-  await loadMysqlCheckConstraints(executor, snapshot, schemasIn, excludeIn)
-
-  return snapshot
-}
-
-async function resolveMysqlSchemaList(
-  executor: SqlExecutor,
-  schemas: string[] | undefined,
-): Promise<string[]> {
-  if (schemas && schemas.length > 0) return schemas
-  const dbRow = await executor.all('SELECT DATABASE() AS db')
-  const db = asString(dbRow[0]?.db)
-  if (!db) {
-    throw new Error(
-      'snapshotSchema: MySQL connection has no current database — pass `schemas` explicitly',
-    )
-  }
-  return [db]
-}
-
-async function loadMysqlTables(
-  executor: SqlExecutor,
-  snapshot: SchemaSnapshot,
-  schemasIn: string,
-  excludeIn: string,
-): Promise<void> {
-  const tables = await executor.all(
-    `SELECT table_schema, table_name
-      FROM information_schema.tables
-      WHERE table_schema IN (${schemasIn})
-        AND table_type = 'BASE TABLE'
-        AND table_name NOT IN (${excludeIn})
-      ORDER BY table_schema, table_name`,
-  )
-  for (const row of tables) {
-    const schema = asString(row.table_schema ?? row.TABLE_SCHEMA)
-    const name = asString(row.table_name ?? row.TABLE_NAME)
-    const contents = snapshot.schemas[schema]
-    if (!contents) continue
-    ensureTable(contents, name)
-  }
-}
-
-async function loadMysqlColumns(
-  executor: SqlExecutor,
-  snapshot: SchemaSnapshot,
-  schemasIn: string,
-  excludeIn: string,
-): Promise<void> {
-  const columns = await executor.all(
-    `SELECT
-        table_schema,
-        table_name,
-        column_name,
-        column_type,
-        is_nullable,
-        column_default,
-        extra,
-        generation_expression
-      FROM information_schema.columns
-      WHERE table_schema IN (${schemasIn})
-        AND table_name NOT IN (${excludeIn})
-      ORDER BY table_schema, table_name, ordinal_position`,
-  )
-  for (const row of columns) {
-    const schema = asString(row.table_schema ?? row.TABLE_SCHEMA)
-    const tableName = asString(row.table_name ?? row.TABLE_NAME)
-    const contents = snapshot.schemas[schema]
-    if (!contents) continue
-    const table = contents.tables[tableName]
-    if (!table) continue
-    table.columns[asString(row.column_name ?? row.COLUMN_NAME)] = buildMysqlColumn(row)
-  }
-}
-
-function buildMysqlColumn(row: Record<string, unknown>): ColumnSnapshot {
-  const generationExpr = asNullableString(row.generation_expression ?? row.GENERATION_EXPRESSION)
-  return {
-    type: asString(row.column_type ?? row.COLUMN_TYPE),
-    nullable: asString(row.is_nullable ?? row.IS_NULLABLE).toUpperCase() === 'YES',
-    default: asNullableString(row.column_default ?? row.COLUMN_DEFAULT),
-    identity: null,
-    extra: asString(row.extra ?? row.EXTRA),
-    generated: generationExpr && generationExpr.length > 0 ? generationExpr : null,
-  }
-}
-
-interface MysqlIndexAggregate {
-  unique: boolean
-  method: string
-  columns: { seq: number; name: string }[]
-}
-
-async function loadMysqlIndexesAndKeys(
-  executor: SqlExecutor,
-  snapshot: SchemaSnapshot,
-  schemasIn: string,
-  excludeIn: string,
-): Promise<void> {
-  const indexRows = await executor.all(
-    `SELECT
-        table_schema,
-        table_name,
-        index_name,
-        non_unique,
-        index_type,
-        seq_in_index,
-        column_name
-      FROM information_schema.statistics
-      WHERE table_schema IN (${schemasIn})
-        AND table_name NOT IN (${excludeIn})
-      ORDER BY table_schema, table_name, index_name, seq_in_index`,
-  )
-
-  const indexAccum = new Map<string, MysqlIndexAggregate>()
-  for (const row of indexRows) {
-    accumulateMysqlIndexRow(row, indexAccum)
-  }
-
-  for (const [key, agg] of indexAccum) {
-    const [schema, tableName, indexName] = splitKey3(key)
-    const table = snapshot.schemas[schema]?.tables[tableName]
-    if (!table) continue
-    applyMysqlIndexAggregate(table, indexName, agg)
-  }
-}
-
-function accumulateMysqlIndexRow(
-  row: Record<string, unknown>,
-  accum: Map<string, MysqlIndexAggregate>,
-): void {
-  const schema = asString(row.table_schema ?? row.TABLE_SCHEMA)
-  const tableName = asString(row.table_name ?? row.TABLE_NAME)
-  const indexName = asString(row.index_name ?? row.INDEX_NAME)
-  const key = `${schema}.${tableName}.${indexName}`
-  let agg = accum.get(key)
-  if (!agg) {
-    agg = {
-      unique: !asBool(row.non_unique ?? row.NON_UNIQUE),
-      method: asString(row.index_type ?? row.INDEX_TYPE),
-      columns: [],
-    }
-    accum.set(key, agg)
-  }
-  agg.columns.push({
-    seq: Number(row.seq_in_index ?? row.SEQ_IN_INDEX ?? 0),
-    name: asString(row.column_name ?? row.COLUMN_NAME),
-  })
-}
-
-function applyMysqlIndexAggregate(
-  table: TableSnapshot,
-  indexName: string,
-  agg: MysqlIndexAggregate,
-): void {
-  const columns = agg.columns.sort((a, b) => a.seq - b.seq).map((c) => c.name)
-  if (indexName === 'PRIMARY') {
-    table.primaryKey = { name: 'PRIMARY', columns }
-    return
-  }
-  if (agg.unique) {
-    table.uniqueConstraints[indexName] = { columns }
-  }
-  table.indexes[indexName] = { columns, unique: agg.unique, method: agg.method }
-}
-
-interface MysqlFkAggregate {
-  schema: string
-  tableName: string
-  name: string
-  refSchema: string
-  refTable: string
-  onUpdate: ForeignKeyAction
-  onDelete: ForeignKeyAction
-  columns: { ord: number; col: string; refCol: string }[]
-}
-
-async function loadMysqlForeignKeys(
-  executor: SqlExecutor,
-  snapshot: SchemaSnapshot,
-  schemasIn: string,
-  excludeIn: string,
-): Promise<void> {
-  const fkRows = await executor.all(
-    `SELECT
-        rc.constraint_schema AS schema_name,
-        rc.table_name AS table_name,
-        rc.constraint_name AS constraint_name,
-        rc.update_rule AS update_rule,
-        rc.delete_rule AS delete_rule,
-        rc.referenced_table_name AS ref_table,
-        kcu.referenced_table_schema AS ref_schema,
-        kcu.column_name AS column_name,
-        kcu.referenced_column_name AS ref_column,
-        kcu.ordinal_position AS ord
-      FROM information_schema.referential_constraints rc
-      JOIN information_schema.key_column_usage kcu
-        ON kcu.constraint_schema = rc.constraint_schema
-       AND kcu.constraint_name = rc.constraint_name
-       AND kcu.table_name = rc.table_name
-      WHERE rc.constraint_schema IN (${schemasIn})
-        AND rc.table_name NOT IN (${excludeIn})
-      ORDER BY rc.constraint_schema, rc.table_name, rc.constraint_name, kcu.ordinal_position`,
-  )
-
-  const fkAccum = new Map<string, MysqlFkAggregate>()
-  for (const row of fkRows) {
-    accumulateMysqlFkRow(row, fkAccum)
-  }
-  for (const agg of fkAccum.values()) {
-    const table = snapshot.schemas[agg.schema]?.tables[agg.tableName]
-    if (!table) continue
-    const sorted = agg.columns.sort((a, b) => a.ord - b.ord)
-    table.foreignKeys[agg.name] = {
-      columns: sorted.map((c) => c.col),
-      referencedSchema: agg.refSchema,
-      referencedTable: agg.refTable,
-      referencedColumns: sorted.map((c) => c.refCol),
-      onUpdate: agg.onUpdate,
-      onDelete: agg.onDelete,
-    }
-  }
-}
-
-function accumulateMysqlFkRow(
-  row: Record<string, unknown>,
-  accum: Map<string, MysqlFkAggregate>,
-): void {
-  const schema = asString(row.schema_name)
-  const tableName = asString(row.table_name)
-  const constraintName = asString(row.constraint_name)
-  const key = `${schema}.${tableName}.${constraintName}`
-  let agg = accum.get(key)
-  if (!agg) {
-    agg = {
-      schema,
-      tableName,
-      name: constraintName,
-      refSchema: asString(row.ref_schema),
-      refTable: asString(row.ref_table),
-      onUpdate: normalizeAction(asString(row.update_rule)),
-      onDelete: normalizeAction(asString(row.delete_rule)),
-      columns: [],
-    }
-    accum.set(key, agg)
-  }
-  agg.columns.push({
-    ord: Number(row.ord),
-    col: asString(row.column_name),
-    refCol: asString(row.ref_column),
-  })
-}
-
-async function loadMysqlCheckConstraints(
-  executor: SqlExecutor,
-  snapshot: SchemaSnapshot,
-  schemasIn: string,
-  excludeIn: string,
-): Promise<void> {
-  let checkRows: Record<string, unknown>[]
-  try {
-    checkRows = await executor.all(
-      `SELECT
-          tc.table_schema,
-          tc.table_name,
-          tc.constraint_name,
-          cc.check_clause
-        FROM information_schema.table_constraints tc
-        JOIN information_schema.check_constraints cc
-          ON cc.constraint_schema = tc.constraint_schema
-         AND cc.constraint_name = tc.constraint_name
-        WHERE tc.constraint_type = 'CHECK'
-          AND tc.table_schema IN (${schemasIn})
-          AND tc.table_name NOT IN (${excludeIn})
-        ORDER BY tc.table_schema, tc.table_name, tc.constraint_name`,
-    )
-  } catch (error) {
-    // information_schema.check_constraints didn't ship until MySQL 8.0.16. Silently swallow
-    // only that specific missing-table case; everything else (permissions, catalog corruption,
-    // network) must surface so callers don't get silently incomplete snapshots.
-    const message = error instanceof Error ? error.message : ''
-    if (
-      /check_constraints/i.test(message) &&
-      /(unknown table|doesn't exist|does not exist|not exist)/i.test(message)
-    ) {
-      return
-    }
-    throw error
-  }
-  for (const row of checkRows) {
-    const schema = asString(row.table_schema ?? row.TABLE_SCHEMA)
-    const tableName = asString(row.table_name ?? row.TABLE_NAME)
-    const constraintName = asString(row.constraint_name ?? row.CONSTRAINT_NAME)
-    const table = snapshot.schemas[schema]?.tables[tableName]
-    if (!table) continue
-    table.checkConstraints[constraintName] = {
-      expression: asString(row.check_clause ?? row.CHECK_CLAUSE),
-    }
-  }
-}
-
-function splitKey3(key: string): [string, string, string] {
-  // Robustly split a 3-part key on the rightmost two dots. Schema/table names cannot contain
-  // dots in practice for our caller, so a simple split works, but be defensive.
-  const idx2 = key.lastIndexOf('.')
-  const idx1 = key.lastIndexOf('.', idx2 - 1)
-  return [key.slice(0, idx1), key.slice(idx1 + 1, idx2), key.slice(idx2 + 1)]
-}
-
-// ──────────────────── Public entry point ────────────────────
-
-/**
- * Captures a structural snapshot of a live database schema by querying its system catalogs.
- *
- * Designed for the Prisma → Drizzle (or any cross-ORM) migration use case:
- * after generating Drizzle migrations and running them against a fresh database, compare
- * the result against a snapshot of the original Prisma-built database. Any difference is
- * a real defect in the Drizzle schema definition (typically a missing explicit name
- * override on a constraint or index) that would cause `markMigrationsApplied` to record
- * a baseline that doesn't actually match the production schema.
- *
- * The snapshot is intentionally strict — names are part of the schema and are not
- * normalized away. Column ordering is the only thing not enforced (columns are keyed
- * by name, not by ordinal position).
- *
- * Supported dialects: PostgreSQL, MySQL, CockroachDB (uses the PostgreSQL path).
- */
-export function snapshotSchema(options: SnapshotSchemaOptions): Promise<SchemaSnapshot> {
-  const { executor, dialect } = options
-  const excludeTables = new Set(options.excludeTables ?? DEFAULT_EXCLUDE_TABLES)
-
-  if (isPgLike(dialect)) {
-    const schemas = options.schemas && options.schemas.length > 0 ? options.schemas : ['public']
-    return snapshotPostgres(executor, schemas, excludeTables, dialect)
-  }
-
-  if (dialect === 'mysql') {
-    return snapshotMysql(executor, options.schemas, excludeTables)
-  }
-
-  return Promise.reject(new Error(`Unsupported dialect for snapshotSchema: "${dialect}"`))
-}
